+++ src/components/badge/types.ts
@@ -1,0 +1,15 @@
+export enum BadgeTypes {
+  block = 'block',
+  info = 'info',
+  cancel = 'cancel',
+  success = 'success',
+  warning = 'warning',
+  error = 'error',
+  custom = 'custom',
+}
+
+export enum BadgeFill {
+  full = 'full',
+  half = 'half',
+  none = 'none',
+}