+++ src/components/notification/script.ts
@@ -1,11 +1,7 @@
-import {
-  Component,
-  Prop,
-  Vue,
-} from 'vue-property-decorator';
+import { Component, Prop, Vue } from 'vue-property-decorator';
 
 import StIcon from '../icon/index.vue';
-import { NOTIFICATION_TYPES } from './types';
+import { NotificationTypes } from './types';
 
 
 @Component({
@@ -24,8 +20,11 @@ export default class StNotification extends Vue {
   @Prop({ type: String, required: true })
   message!: string;
 
-  @Prop({ type: String, default: NOTIFICATION_TYPES.INFO })
-  type!: NOTIFICATION_TYPES;
+  @Prop({ type: String, default: NotificationTypes.info })
+  type!: NotificationTypes;
+
+  @Prop(String)
+  customType!: string;
 
   @Prop({ type: Boolean, default: true })
   closeable!: boolean;
@@ -39,6 +38,10 @@ export default class StNotification extends Vue {
   timer: ReturnType<typeof setTimeout> = 0;
 
   visible = false;
+
+  get typeName(): string {
+    return this.type === NotificationTypes.custom ? this.customType : this.type;
+  }
 
   mounted() {
     this.$nextTick(this.initialize);